### 2021-03-26 #52 @lecfab
Dear all,
This is a replication written with Lionel Tabourier and Maximilien Danisch. It includes the implementation and test of various graph algorithms. The code is mainly in C++ and we suggest Nicolas P. Rougier as the editor.

Best regards,
Fabrice Lécuyer
