### 2021-04-07 #52 @lecfab
Hello @rougier, thanks for your reply.
After looking into Rescience reviewers, we suggest : @EHadoux and @ozancaglayan.
For people outside of Rescience whose research interests are closer to our replication, we suggest Antoine Amarilli (https://a3nm.net, a3nm@a3nm.net) or Stefano Leucci (https://www.stefanoleucci.com, stefano.leucci@univaq.it).
