### 2021-04-12 #52 @ozancaglayan
A preliminary review for the **paper only**, comments on the code and the reproducibility of the experimental pipeline will follow:

This paper replicates the results provided by the paper "Speedup Graph Processing by Graph Ordering by Hao Wei,
Jeffrey Xu Yu, Can Lu, and Xuemin Lin, published in Proceedings of SIGMOD 2016.". The paper is very well written and is a pleasure to read, offering thorough analyses across different ordering methods explored in the original paper. The authors also
incorporate a 9th "Random" ordering as the least favorable i.e. ~worst-case ordering. I particularly liked the way
results are compared in Figure 6. I have very minor questions and suggestions in what follows, none of them is critical:

**Questions:**

- Table.3: I have a hard time to decipher the columns here. What is `-r` and `-mr` in columns 2, 4, and 5?
  What does it mean that the "percentage columns respectively show the proportion of data that had to be retrieved beyond
  L1, L2 and L3."? Although the text clarifies these, I think it would be good to rework the table caption a bit.

- I was curious on how Fig.5 would look if the x-axis groups the orderings and the bars reflect the datasets. Maybe the
  plot will be easier to read with less oscillation of the bars? Just thinking loudly, never mind if this doesn't make sense.

**Other remarks:**

`Page 3. shortest paths`
  - What does `m` denote here? It would be good to have it defined at the end of 2.1's beginning paragraph.
  - near the end: "this algorithms" -> "this algorithm"

`Page 5. Energy formulations`
  - I understand that \pi refers to a particular permutation of nodes but what does \pi_u and \pi_v denote here with
respect to nodes u and v? In page 6 (RCM), it is said that "\pi_u is the index of node u". Could you clarify these in the text?

- How does the given energy function kick in for the first equations in page 5? Should that read
`e(abs(\pi_u - \pi_v))` with e() denoting the energy function, for example?

`Page 10 - line 3.`
  - "In a fourth of the experiments" -> In one fourth? or fourth set of?
