### 2021-04-12 #52 @ozancaglayan
## Running the code

### Steps followed

The compilation of the code is straight-forward and pretty quick:

$ git clone --recurse-submodules https://github.com/lecfab/rescience-gorder.git
$ cd rescience-gorder/src
$ make
```

I tried to run the benchmark script on a Ubuntu 20.04.2 laptop booted with the Linux kernel version 5.8.0.

$ ./run-benchmark.sh
Cache sizes for each processor of this machine:
  Processor 0:	L0-Data: 32K 	L1-Instruction: 32K 	L2-Unified: 256K 	L3-Unified: 6144K
  Processor 1:	L0-Data: 32K 	L1-Instruction: 32K 	L2-Unified: 256K 	L3-Unified: 6144K
  Processor 2:	L0-Data: 32K 	L1-Instruction: 32K 	L2-Unified: 256K 	L3-Unified: 6144K
  Processor 3:	L0-Data: 32K 	L1-Instruction: 32K 	L2-Unified: 256K 	L3-Unified: 6144K
  Processor 4:	L0-Data: 32K 	L1-Instruction: 32K 	L2-Unified: 256K 	L3-Unified: 6144K
  Processor 5:	L0-Data: 32K 	L1-Instruction: 32K 	L2-Unified: 256K 	L3-Unified: 6144K
  Processor 6:	L0-Data: 32K 	L1-Instruction: 32K 	L2-Unified: 256K 	L3-Unified: 6144K
  Processor 7:	L0-Data: 32K 	L1-Instruction: 32K 	L2-Unified: 256K 	L3-Unified: 6144K
Warning: impossible to measure cache in unknown machines; please install ocperf (see README) and edit this script to get cache measurements.
Experiments will soon start... (runtime will be measured but not cache-miss)
Results will be found in ../results/r4949
<output stripped>
```

#### Remarks

- [x] The script succesfully produces runtime measurements related files and plots.
- [ ] However, the script warns that `ocperf` should be installed and points us to the README file. However, there are no instructions in the README file for `ocperf` installation. In fact, the `ocperf` script comes from the `pmu-tools` submodule which
is already checked out. I think the warning is misleading because the problem is about **not passing a command-line
argument to the script to identify the machine configuration**.

The only configuration defined is `mesu`, which apparently is a cluster configuration that the authors used. I gave that a try as well:

$ ./run-benchmark.sh mesu
...
../pmu-tools/ocperf stat -e task-clock,cpu-cycles,instructions,L1-dcache-loads,L1-dcache-load-misses,LLC-loads,LLC-load-misses,cycle_activity.cycles_l1d_pending,cycle_activity.cycles_l2_pending -o ../results/r2118/perf-epinion-original-nq.txt ./benchmark ../datasets/edgelist-epinion-75k-508k.txt -a nq -o ../results/r2118/time-epinion-original-nq.txt -l 10
Cannot run perf
```

#### Remarks

- [ ] I now realize that the `perf` util should be installed on the system. Could you add a simple check in the script
to detect whether `perf` is installed or not and tell the user to install it using the distribution's package manager? For my kernel version, the following command did the trick: `apt install linux-tools-5.8.0-48-generic linux-cloud-tools-5.8.0-48-generic`
- [ ] It may actually be much more clearer if you split runtime measurements and cache measurements into two different scripts? What do you think on this?

Let's gave the script another try after installing `perf`:

$ ./run-benchmark.sh mesu
../pmu-tools/ocperf stat -e task-clock,cpu-cycles,instructions,L1-dcache-loads,L1-dcache-load-misses,LLC-loads,LLC-load-misses,cycle_activity.cycles_l1d_pending,cycle_activity.cycles_l2_pending -o ../results/r2959/perf-epinion-original-nq.txt ./benchmark ../datasets/edgelist-epinion-75k-508k.txt -a nq -o ../results/r2959/time-epinion-original-nq.txt -l 10
event syntax error: '..oad-misses,cycle_activity.cycles_l1d_pending,cycle_act..'
                                  \___ parser error
Run 'perf list' for a list of valid events

 Usage: perf stat [<options>] [<command>]
...
...
Warning: dataset pokec does not exist.
Warning: dataset flickr does not exist.
Warning: dataset livejournal does not exist.
Warning: dataset wiki does not exist.
Warning: dataset gplus does not exist.
Warning: dataset pldarc does not exist.
Warning: dataset twitter does not exist.
Warning: dataset sdarc does not exist.

Creating figures for orders comparison
Relatve runtime and ranking frequency
Warning: 9 datasets miss Gorder on NQ
Warning: 9 datasets miss Gorder on BFS
Warning: 9 datasets miss Gorder on DFS
Warning: 9 datasets miss Gorder on SCC
Warning: 9 datasets miss Gorder on SP
Warning: 9 datasets miss Gorder on PR
Warning: 9 datasets miss Gorder on DS
Warning: 9 datasets miss Gorder on Kcore
Warning: 9 datasets miss Gorder on Diam
Worse score  1
Image has been saved in /home/ozan/git/rescience-gorder/results/r2959/img-speedup.pdf
Image has been saved in /home/ozan/git/rescience-gorder/results/r2959/img-ranking.pdf
CPU execution and cache stall of Gorder and Original on sdarc
Error with /home/ozan/git/rescience-gorder/results/r2959/perf-sdarc-gorder-nq.txt
Error with /home/ozan/git/rescience-gorder/results/r2959/perf-sdarc-gorder-nq.txt
Error with /home/ozan/git/rescience-gorder/results/r2959/perf-sdarc-original-nq.txt
Error with /home/ozan/git/rescience-gorder/results/r2959/perf-sdarc-original-nq.txt
nq [] []
  File "../results/gorder-cache-bars.py", line 31, in <module>
    gdata_stall.append(gperf[1] / operf[0])
Cache-miss rates for pr on sdarc
Error with /home/ozan/git/rescience-gorder/results/r2959/perf-sdarc-original-pr.txt
Error with /home/ozan/git/rescience-gorder/results/r2959/perf-sdarc-original-pr.txt
Error with /home/ozan/git/rescience-gorder/results/r2959/perf-sdarc-original-pr.txt
Error with /home/ozan/git/rescience-gorder/results/r2959/perf-sdarc-original-pr.txt
  File "../results/gorder-cache-table.py", line 47, in <module>
    billions(specs[0]),         # number of L1 references
```

#### Remarks

Now apparently all the `perf` calls are failing because of the `mesu` config not being appropriate for the CPUs that I have.
They fail on both Intel Xeon Silver 4116 and Intel Skylake i5-8250 CPU. This further leads to errors in the python scripts
probably because there are no results produced for cache statistics.

- [ ] Is it difficult to provide a sane set of defaults for `perf` that'll work on most of the modern machines out there and
  will produce a subset of the results at least?
- [ ] After a brute force trial-and-error, I removed the non-supported `cycle_activity.*` events from perf calls and `perf`
began to work but only after setting a `procfs` flag to `-1` to allow regular users profiling **or** running the script as root. Could you add this into README as well? Although this created some cache-related stats as well, the python scripts still failed probably because of the missing perf events.
- [ ] Even the problematic events were supported on my machine and everything was working flawlessly, I see that `gorder-cache-bars.py` and `gorder-acche-table.py` actually computes **the plots on the `sdarc` dataset.** However, we are only given the `epinion` dataset in this repository.  I though that at least we will be able to produce all sets of plots/graphics from the paper but only on the `epinion` dataset. Could you clarify this? If `epinion` doesn't make sense for some experiments simply it is a quite small/simple dataset, could you document where and how to get the other datasets in the `README.md` file?
- [ ] The README file is self-contained as it explains all the codes/scripts provided. However, I think the authors should add a particular section **on top of the README file** to explain the exact steps that are required to reproduce the results and the plots in their paper. Right now, these information are scattered into different sections of the README.md file.

**NOTE:** After looking through Google, I understand that the `cycle_activity.*` events are not supported on new Intel CPUs. The list of events for Skylake is [here](https://download.01.org/perfmon/index/skylake.html). On some other posts, I see that these events may not be reliable, i.e. [1](https://community.intel.com/t5/Software-Tuning-Performance/difference-between-STALLS-L1D-PENDING-and-CYCLES-L1D-PENDING/td-p/996943), [2](https://community.intel.com/t5/Software-Tuning-Performance/Are-CYCLE-ACTIVITY-STALLS-LDM-PENDING-and-CYCLE-ACTIVITY-CYCLES/td-p/1051707) So I wonder, whether there are some other events there, that are more robust and supported on newer CPUs? I should note that I am by no means an expert on these profiling stuff.

## Comparing the results

I don't know what kind of information does this give but, I ran the benchmark twice (1) on my laptop and (2) on a server. In both cases, the system load was stable and low, no desktop environment or other processes were running. I created a plot that shows all 4 runs (2x laptop, 2x server) side by side compared to the plot in the paper, on the `epinion` dataset:

![gorder](https://user-images.githubusercontent.com/330946/114523048-9bc8ae80-9c4c-11eb-9886-d2f4ec45f942.png)

To me, the results look almost compatible across the board. Here are some questions:
- Do you plot the mean or the median of 10 repetition runtimes? Is it possible that some of the runs are outliers because of dataset loading into RAM?
- It's curious that for NQ, the server with more CPUs slowed down the runtime of three algos. Are the implementations thread/process aware? For the paper, those algos apparently perform as good as Gorder since we don't have bars in the plot, right?
- In overall, `ChDFS, InDegSort and RCM` seem to be the ones that are not that stable when compared to the paper, in terms of conclusions.

(I know that, this is just on `epinion` but it still can show some interesting insights to us.)
