### 2021-04-13 #52 @EHadoux
Here is a review of the paper, the review of running the code will come shortly.

____
This paper is the reproduction of _Speedup Graph Processing by Graph Ordering_ by Wei et. al. In the original paper, the authors proposed an algorithm and graph agnostic node ordering procedure allowing for the reduction of CPU cache miss. The reproduction paper reproduces the experiment section on the same 9 algorithms and 8 datasets. It also contains an additional, smaller dataset.

## Review
The paper reads nicely, addressing each point in turn in a non-confusing way, sometimes even adding more information than what the original paper contains. I appreciate that the authors have made their best efforts to reproduce the original implementation and, when not enough information was given, provided a close alternative instead of dropping it out altogether.

Comparing all the results, the authors draw the same as in the original paper, albeit to a milder extent concerning the performance of the Gorder method. It would have been interesting to discuss a bit more (if possible) the potential reasons for this difference.

## Remarks
- Each chart in Figure 5 could benefit from having vertical lines or different backgrounds for each graph, it’s sometimes a bit hard to see at a glance where one graph (i.e. a group of bars) ends and another one starts.
- Comparing Table 2 in the reproduction and Table 9 in the original paper, most lines make sense given the explanations (e.g. SlashBurn), although I am surprised that most computation times are significantly faster in the reproduction with a less performant hardware. I don’t think it’s a problem in itself though given the objective of the paper.

## Questions
- Keeping in mind what is being said about the hardware (i.e. more performant in the original paper), does it mean that there is a speed/memory ratio below which this optimisation makes the improvement marginal?
- Comparing Figure 1 in both papers as well as both Table 3, it looks like the reproduction implementation has less cache misses overall, even before reordering. Would it mean that a poorer initial implementation would benefit more from this optimisation than an already efficient one?
- In the « limits of the visualisation », the paper mentions that the original paper « hides precise information the a runtime exceeds 1.5 times the runtime of Gorder ». Is there for each of them a permutation of rankings that would make the overall ranking exactly the same between the two papers? I’m not sure it is interesting though as this only concerns the slowest methods we don’t want to use anyway.
