### 2021-04-16 #52 @EHadoux
I'll add my review of the code most probably this evening.
