### 2021-04-17 #52 @EHadoux
## Review of the code and the instructions

### General review
Past the very small number of hurdles (see below) the whole process is very smooth. The automatic generation of the charts is **greatly** appreciated.

I am confident that the results provided in the paper are a true depiction of the outputs of the code provided with it.

### Specific remark
> Download datasets in datasets/dl/ and use datasets/normalise.py to match the edgelist format. Note that epinion is readily available for quick testing.

Is the epinion file that's already there the output of ```normalise.py``` for this dataset? I assume yes because I can't use it as an input. Also, giving the direct links to the datasets would help. I found a link to a Flickr dataset http://konect.cc/networks/flickr-growth/ but I am not 100% sure it is the right one.

* ```make``` doesn't work on Mac with Clang because it seems that concepts aren't available. Installing gcc through ```brew install gcc``` and modifying the Makefile solved the problem. Maybe adding these instructions or removing the concepts from the code if they are not mandatory.

* The scripts require some Python packages (e.g. Matplotlib), providing a ```requirement.txt``` that we can just ```pip install``` would be good.

> ```$ ./run-window.sh [REPEAT]``` tests different window sizes for Gorder on Flickr dataset
The script actually runs on epinion, it needs to be slightly modified to run on Flickr. However the modification to make is already in the script, commented. It is extremely obvious how to change it, but one would need to open the source code.

### Using the provided scripts only
#### run-window
I used 1 as a parameter unlike 100 as the authors reported due to the time it takes to run it, and stopped at a window size of 16384 (2^14 < 2^20 in the paper). You can see below the output generated. We can see the downward trend, combined with the chart on epinion (downwards then upwards), I have confidence the chart on Flickr would have looked like the one reported in the paper.

![tune-w-flickr-pr](https://user-images.githubusercontent.com/1655646/115118595-d8bdd980-9f9b-11eb-9ffc-3418b39c8605.png)
![tune-w-epinion-pr](https://user-images.githubusercontent.com/1655646/115118599-de1b2400-9f9b-11eb-8914-f9be3909071d.png)

#### run-annealing

It seems that the paper does not state which dataset has been used to generate Figure 3. I assumed it was epinion as it is the default setting in ```run-annealing.sh```.

Although the units on the axes are different, the shapes for both graphs are very similar to the reported one.

![Capture d’écran 2021-04-17 à 16 49 37](https://user-images.githubusercontent.com/1655646/115118805-f475af80-9f9c-11eb-8d81-26053f187080.png)

#### run-benchmark
I cannot run it on Mac as the script uses /proc/cpuinfo that does not exist on Mac. I'll run it if the authors provide an alternative script that would cater for Mac.

I have yet to run the ```./benchmark``` but glancing at the code, I am confident that what has been committed has been generated by the C++ code, if the same Python/Bash scripts have been used to collate the results.
