### 2021-04-19 #52 @lecfab
Here are the first answers to your interesting remarks, and the corrections i made to the paper. The new version is not yet uploaded on github. Feel free to react to any of my comments below!

I will work on the code and your replications as soon as possible.

# First updates after review

## Remarks
### Ozan Caglayan's comments

1. > Could bars in figure 5 be grouped by order instead of by dataset?

    Nice suggestion, I tried it and it turns out to be helpful to compare the efficiency of the orderings. It may be misleading though, because the time scale is not the same for different datasets. I will add this figure in supplementary material: [img-speedup-grouped.pdf](https://github.com/ReScience/submissions/files/6353197/img-speedup-grouped.pdf)

1. > In simulated annealing, how is the variation of energy e related to the total energy EE?

    If indices of u and v are swapped, we could write (for MinLA):<br>
![$$e=\sum_{w\in N_u} (|\pi_v-\pi_w| - |\pi_u-\pi_w|)  +  \sum_{w\in N_v} (|\pi_u-\pi_w| - |\pi_v-\pi_w|)$$](https://www.bruot.org/tex2img/media/IfMxFRiXVsu4a9Wv2wht1ptWrmpbtG8ySxhxTyJ2MtSL/tex2img_equation.svg)<br>
The first sum is the energy variation for u, and the second is for v. In each sum, the first term adds the energy due to the new index and substracts the energy due to the old index. So e represents the change in the contribution of u and v in EE. Is this any clearer?

1. > Runtime and cache measurements could be separated

    Indeed it would clearer with separate scripts. However, the full experiment takes about a week, so we thought it was better to run them only once, and measure both time and cache-miss rates.

1. > Do you plot the mean or the median of 10 repetition runtimes?

    We plot the sum of 10 repetitions (which corresponds to averaging).

1. > Can some of the 10 repetitions be outliers because they include dataset loading?

    The algorithms are repeated 10 times _after_ loading the graph in RAM, so this should not be the issue. Yet, we had many issues with fluctuations: on our machines, repeating the same procedure could have up to 20% of time variability (due to process competition, cache or RAM sharing, multithreading...). As we are not specialists for these issues, we ended up running everything on an isolated node (called _mesu_ indeed!).

    Note that _epinion_ is even more subject to fluctuations because it is small and algorithms only take a fraction of a second.

1. > It's curious that for NQ, the server with more CPUs slowed down the runtime of three algos. Are the implementations thread/process aware? For the paper, those algos apparently perform as good as Gorder since we don't have bars in the plot, right?

    It means that the runtimes are very similar (under 10% of variation). Note that _epinion_ is probably too small to see the effect of wise ordering. On my machine, NQ is also the algorithm with most computation time (figure 1) and less cache stall, so the fluctuations on this small dataset may be more significant than the speedup due to ordering.

    Our code uses only one thread but is not protected from other threads, and that's maybe a reason why we had to use an isolated cluster to avoid heavy fluctuations. I believe the 6 bars above 2.2 are a sign of such unpredictable fluctuations.

1. > In overall, ChDFS, InDegSort and RCM seem to be the ones that are not that stable when compared to the paper, in terms of conclusions.

    I'm not sure: on your results, we see that these 3 orderings have small or invisible bars in either direction. It means that they are quite close to the performance of gorder and sometimes better. This sort of matches my results, though I think we reach the limit for _epinion_ in term of analysis!

### Emmanuel Hadoux's comments

11. > Why are ordering times so different between the original work and the replication? (table 2)

    While coding MinLA, MinLoga, Slashburn and LDG, we had to make a lot of choices of assumptions because the original paper did not explain much. (For instance, our Slashburn is a very simplified version, quite similar to Kcore). For this reason, the ordering time cannot be compared across the two papers. The remaining orderings (RCM, Deg, DFS) are quite standard so comparing is reasonable. Our Gorder was heavily based on the code provided by the original authors, so the comparison stands as well. With this restriction, our hardware is 2 to 5 times slower, which we think makes sense.

1. > Is there a speed/memory ratio below which this optimisation makes the improvement marginal?

    On most architectures, part of the data is physically closer to the processor and thus faster to retrieve. I'm really not an expert in hardware, but I would think that if processor speed is increased, most of the time will be spent in data retrieval (cache stall), which in figure 1 would make the grey bars tiny. In that case, cache optimisation is even more important. On the contrary, if the CPU slow, big amounts of distant data can be fetched while the CPU makes an operation, so cache optimisation is not meaningful.

1. > Comparing Figure 1 in both papers as well as both Table 3, it looks like the reproduction implementation has less cache misses overall

    I think the rates across machines cannot be compared without knowing the details of hardware architecture.

1. > In the « limits of the visualisation », is there a permutation of rankings beyond x1.5 that would make the overall ranking exactly the same between the two papers?

    No, indeed the approximation of cutting everything above x1.5 only affects the poor performers, and we see that good performers do not rank equally in both papers. To try and be fair, we gave the same rank to all the out-of-limits orderings (if there are 10 orderings and 6 of them are beyond x1.5, we all rank them fifth) so that Fig.6 is deterministic.

Major modifications are in brown in the updated paper of the repository.

* [x] First page: **add abstract** and reference to replicated paper
* [x] Method: define $m$: number of edges of a graph
* [x] Method: define $\pi_u$: index of node u in arrangement pi
* [x] RCM: explain $\pi_u$ in the definition of the ordering
* [x] indicate which dataset is used for simulated annealing: _epinion_
* [x] Gorder: **re-explain window tuning**
* [x] figure 5: add to supplementary material a **version grouped by ordering**
* [x] figure 5: add absolute time information; update caption
* [x] figure 5: separate blocks of bars with gray highlighting
* [x] table 2: update caption to explain time differences
* [x] table 3: the column names have been copied from the original paper to facilitate comparison; update the caption to make them clearer
* [x] section 3.4: **more discussion** about the difference of performance between original and replication
* [x] section 3.5: proper **definition of cache-miss**
* [x] typoes

* [x] provide datasets link in readme > datasets > click on a NAME

### Ozan Caglayan's comments
* [x] indicate how to install perf in readme > tools and versions
* [x] clean warning messages when several datasets are missing
* [x] check if perf is installed and propose install command otherwise
* [x] change benchmark.sh warning message
* [x] change figure 1 and table 3 generator when datasets are missing
* [x] add the procfs -1 setting into readme > tools and versions
* [x] benchmark: replace `MACHINE` with `basic` and `advanced` modes to activate different metrics. Basic is enough for Table 3 while Figure 1 requires advanced

21. > Add a section on top in the readme with the exact steps required to reproduce the paper

    I intended the _Quick guide_ section to achieve this, could you please suggest what to add in it to complete it?

### Emmanuel Hadoux's comments
* [x] remove `-fconcepts` in makefile (seems harmless)
* [x] add `pip3 install matplotlib numpy` to readme

31. > I used 1 as a parameter unlike 100 as the authors reported due to the time it takes to run it, and stopped at a window size of 16384 (2^14 < 2^20 in the paper). You can see below the output generated. We can see the downward trend, combined with the chart on epinion (downwards then upwards), I have confidence the chart on Flickr would have looked like the one reported in the paper.

    If possible time-wise, it would be interested to test with at least 3 repetitions (on _flickr_), to temper sudden fluctuations. I thik they are responsible for the big peaks in the figure.

1.  > I cannot run it on Mac as the script uses /proc/cpuinfo that does not exist on Mac.

    Could you please provide the error message that this generates? A missing processor info should not stop the script, but i'm not familiar with mac (maybe some shell functions are different)
