### 2021-04-21 #52 @lecfab
@EHadoux: sorry for the mac-related problems! For `run-benchmark.sh`, i think it works if you remove all the _display cache size information_ section (lines 4 to 21).
If you want to use other datasets, there links are given in the replication paper (in table 1). I will add them to the readme though.
