### 2021-04-21 #52 @EHadoux
> @EHadoux: sorry for the mac-related problems! For `run-benchmark.sh`, i think it works if you remove all the _display cache size information_ section (lines 4 to 21).
>
> If you want to use other datasets, there links are given in the replication paper (in table 1). I will add them to the readme though.

Cheers for that, I'll check it soon.
For the concepts bits (I've seen you've put a ?), just in case it wasn't clear, you compile with the -fconcepts flag which is not available on Clang. If you don't use the notion of concepts in your code (I haven't checked), you can probably remove the flag altogether and not have to bother with instructions for Mac users.
