+++ results/gorder-speedup.py
@@ -12,7 +12,9 @@ print("Relatve runtime and ranking frequency")
 # extract data from time-measurement files
 # ----------------------------------------
 data = []
+missing_datasets = []
 for dataset in datasets:
+    dataset_seen = False
     order_data = []
     for order in orders:
         algo_data = [0] * A
@@ -27,78 +29,15 @@ for dataset in datasets:
                         algo_name, time = l.split("\n")[0].split("\t")
                         time = int(time)
                     if algo_name == algo:
+                        dataset_seen = True
                         algo_data[algos.index(algo_name)] += time / 1000 # ms to s
             except IOError as e:
-                # print(e)
                 pass
         order_data.append(algo_data)
     data.append(order_data)
+    if not dataset_seen: missing_datasets.append(dataset)
 
-
-# ----------------------------------------
-# -- draw figure of runtime comparison ---
-# ----------------------------------------
-fig, ax = plt.subplots(5,2, figsize=(210/25,297/25), sharex='col', sharey='row')
-fig.tight_layout(pad=0, w_pad=0, h_pad=.4)
-bar_width = .092  # the width of each little bar
-tick_min, tick_max, tick_step = -.3, 1.2, .2 # y ticks around Gorder result
-tick_range = range(int(tick_min / tick_step) -1, int(tick_max / tick_step) +1)
-# plt.suptitle("Relative speedup of Gorder")
-
-worse_score = 0
-for a in range(A):
-    # ----- choose subplot for each algorithm -----
-    algo = algos[a]
-    aa = a+1 if a > 0 else a
-    sub = aa // 2, aa % 2
-    # print(algo, sub)
-
-    # ----- print title and axis -----
-    ax[sub].set_title('{}'.format(algo_names.get(algo, algo.capitalize())), y=0) #y=-0.15)
-
-    xticks = np.arange(D)  # x labels locations
-    ax[sub].set_xticks(xticks)
-    ax[sub].set_xticklabels(datasets, rotation=45) #, y=-.1)
-    ax[sub].set_xlim(-bar_width*(O+1)/2, (D-1)+bar_width*O/2)
-    # ax[sub].set_ylabel('Relative duration to Gorder')
-    ax[sub].set_yticks([tick_step*i for i in tick_range])
-    ax[sub].set_yticklabels([int(tick_step*(i+1/tick_step)*10)/10 for i in tick_range])
-    ax[sub].set_ylim(tick_min, tick_max)
-    ax[sub].axhline(y = 0, color = 'black', linestyle = '-')
-
-    # ----- find Gorder (reference time) for every dataset ------
-    go = orders.index("gorder")
-    reference = []
-    c = 0
-    for d in range(D):
-        reference.append(data[d][go][a])
-        if reference[d] == 0:
-            reference[d] = 1
-            c += 1
-    if c > 0: print("Warning:", c, "datasets miss Gorder on", algo_names[algos[a]])
-
-    # ----- create histogram bars for each order -----
-    for o in range(O):
-        if orders[o] == "gorder": continue
-        performances = [data[d][o][a] / reference[d] -1 for d in range(D)]
-        performances = [(p if p > -1 else 0) for p in performances]
-        offset = (o - O/2 + 1)*bar_width
-        ax[sub].bar(xticks + offset, performances, bar_width, label=order_names[orders[o]])
-        worse_score = max(worse_score, max(performances)+1)
-
-
-# ----- put legend in the empty subplot [0,1] -----
-legend_lines, legend_labels = ax[0,0].get_legend_handles_labels()
-ax[0,1].axis('off')
-ax[0,1].legend(legend_lines, legend_labels,
[…]
