### 2021-04-23 1c4d40b
Update benchmark, cache.py, readme for unavailable metrics
