### 2021-05-03 #52 @rougier
@lecfab Congratulations, your paper has been accepted! I'll edit it soon and come back to you.
