+++ results/annealing/tune-minla.js
@@ -1,0 +1,84 @@
+data = [
+['MinLA', 'steps', 'kB', 'min', 'max', 'median', 'avg'],
+[18971,6.47423e-12,1.03778e+10,1.03778e+10,1.03778e+10,1.03778e+10],
+[18971,4.05874e-09,1.03778e+10,1.03778e+10,1.03778e+10,1.03778e+10],
+[18971,2.54445e-06,1.03778e+10,1.03778e+10,1.03778e+10,1.03778e+10],
+[18971,0.00159513,1.03778e+10,1.03778e+10,1.03778e+10,1.03778e+10],
+[18971,1,1.03778e+10,1.03778e+10,1.03778e+10,1.03778e+10],
+[18971,626.907,1.03776e+10,1.03776e+10,1.03776e+10,1.03776e+10],
+[18971,393012,1.05978e+10,1.05978e+10,1.05978e+10,1.05978e+10],
+[18971,2.46382e+08,1.05978e+10,1.05978e+10,1.05978e+10,1.05978e+10],
+[18971,1.54458e+11,1.05978e+10,1.05978e+10,1.05978e+10,1.05978e+10],
+[36446,6.47423e-12,1.00049e+10,1.00049e+10,1.00049e+10,1.00049e+10],
+[36446,4.05874e-09,1.00049e+10,1.00049e+10,1.00049e+10,1.00049e+10],
+[36446,2.54445e-06,1.00049e+10,1.00049e+10,1.00049e+10,1.00049e+10],
+[36446,0.00159513,1.00049e+10,1.00049e+10,1.00049e+10,1.00049e+10],
+[36446,1,1.00049e+10,1.00049e+10,1.00049e+10,1.00049e+10],
+[36446,626.907,1.00042e+10,1.00042e+10,1.00042e+10,1.00042e+10],
+[36446,393012,1.05978e+10,1.05978e+10,1.05978e+10,1.05978e+10],
+[36446,2.46382e+08,1.05978e+10,1.05978e+10,1.05978e+10,1.05978e+10],
+[36446,1.54458e+11,1.05978e+10,1.05978e+10,1.05978e+10,1.05978e+10],
+[70016,6.47423e-12,9.64209e+09,9.64209e+09,9.64209e+09,9.64209e+09],
+[70016,4.05874e-09,9.64209e+09,9.64209e+09,9.64209e+09,9.64209e+09],
+[70016,2.54445e-06,9.64209e+09,9.64209e+09,9.64209e+09,9.64209e+09],
+[70016,0.00159513,9.64209e+09,9.64209e+09,9.64209e+09,9.64209e+09],
+[70016,1,9.64209e+09,9.64209e+09,9.64209e+09,9.64209e+09],
+[70016,626.907,9.64458e+09,9.64458e+09,9.64458e+09,9.64458e+09],
+[70016,393012,1.05977e+10,1.05977e+10,1.05977e+10,1.05977e+10],
+[70016,2.46382e+08,1.05978e+10,1.05978e+10,1.05978e+10,1.05978e+10],
+[70016,1.54458e+11,1.05978e+10,1.05978e+10,1.05978e+10,1.05978e+10],
+[134507,6.47423e-12,9.14325e+09,9.14325e+09,9.14325e+09,9.14325e+09],
+[134507,4.05874e-09,9.14325e+09,9.14325e+09,9.14325e+09,9.14325e+09],
+[134507,2.54445e-06,9.14325e+09,9.14325e+09,9.14325e+09,9.14325e+09],
+[134507,0.00159513,9.14325e+09,9.14325e+09,9.14325e+09,9.14325e+09],
[…]
