### 2021-05-18 #52 @lecfab
https://github.com/lecfab/rescience-gorder/tree/main/paper

I filled the files, but could not compile them into a pdf because of the following errors. But i presume it will work on your side after completing the last pieces of information in metadata.yaml.

  <summary>make</summary>
  latexmk -pdf -pdflatex="xelatex -interaction=nonstopmode" -use-make article.tex

Latexmk: This is Latexmk, John Collins, 26 Dec. 2019, version: 4.67.

Latexmk: applying rule 'pdflatex'...

Rule 'pdflatex': The following rules & subrules became out-of-date:      'pdflatex'

------------

Run number 1 of rule 'pdflatex'

------------

Running 'xelatex -interaction=nonstopmode  -recorder  "article.tex"'

------------

sh: 1: xelatex: not found

Latexmk: fls file doesn't appear to have been made.

Latexmk: Errors, so I did not complete making targets

      Refer to 'article.log' for details

----------------------

This message may duplicate earlier message.

Latexmk: Failure in processing file 'article.tex':

   (Pdf)LaTeX didn't generate the expected log file 'article.log'

----------------------

Latexmk: Use the -f option to force complete processing,

 unless error was exceeding maximum runs, or warnings treated as errors.

make: *** [Makefile:30 : article.pdf] Erreur 12
