### 2021-05-31 #52 @rougier
I forgot to make a PR to your repo with the publication changes. Will do in a few minutes.
